Switch from Tailwind to MUI, page routes to app routes
We are going to use MUI for the UI on this project, so I've added some basic setup and there's no real content yet.

I also switched from the old Nextjs `pages` routing to the new `app` routing because it looks a lot more ergonomic and similar to Sveltekit which I'm used to. I set it up with MUI and some layouts and routes. Everything inside the `(public)` group is meant to be pages like help/privacy/home page/about etc, where as anything in `(protected)` is user specific routes, like their dashboard, projects, reports etc. We should be able to protect routes that require the user to be logged in with it setup like this (I'm not sure how, but this is what you might do in Sveltekit so I assume there are similar ways to handle this).

Because `app` route special files are react server components by default they don't have access to things like `useEffect` and `useContext`, and MUI/Emotion don't support server components yet so any file that uses MUI has a `'use client';` directive in it which opts-out of server components for that file.